+++ crates/alint-output/src/junit.rs
@@ -248,6 +248,7 @@ mod tests {
                     column: Some(4),
                     is_note: false,
                     baseline_key: None,
+                    is_fixable: false,
                 }],
             )],
         };
@@ -306,6 +307,7 @@ mod tests {
                     column: None,
                     is_note: false,
                     baseline_key: None,
+                    is_fixable: false,
                 }],
             )],
         };
@@ -329,6 +331,7 @@ mod tests {
                     column: None,
                     is_note: false,
                     baseline_key: None,
+                    is_fixable: false,
                 }],
             )],
         };
@@ -350,6 +353,7 @@ mod tests {
                     column: None,
                     is_note: false,
                     baseline_key: None,
+                    is_fixable: false,
                 }],
             )],
         };